Fold broker negotiation modules into crate roots
