--- litebox_broker_host/src/negotiate.rs
@@ -1,578 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-use core::fmt;
-
-use litebox_broker_core::{BrokerAssociation, BrokerCore, BrokerError, CallerCredential};
-use litebox_broker_protocol::{
-    AddEventResponse, BrokerRequest, BrokerResponse, CoreRequest, CoreResponse,
-    CreateEventResponse, ErrorCode, EventRequest, EventResponse, HostControlChannel,
-    PeerCredential, ProtocolVersion, ReceivedBrokerRequest, WaitEventResponse,
-};
-
-use crate::error::{BrokerHostError, Result as HostResult};
-
-/// Protocol version this broker host implementation supports.
-pub const HOST_PROTOCOL_VERSION: ProtocolVersion = ProtocolVersion::new(0, 2);
-
-/// Serves one broker connection over the provided connected control channel.
-pub fn serve_connection<T>(
-    core: &mut BrokerCore,
-    channel: &mut T,
-) -> HostResult<ConnectionTermination, T::Error>
-where
-    T: HostControlChannel,
-{
-    let peer_credential = channel
-        .peer_credential()
-        .map_err(BrokerHostError::Channel)?;
-    let caller_credential = caller_credential_from_peer(peer_credential)
-        .map_err(|()| BrokerHostError::AssociationSetup)?;
-    let association = core
-        .create_association(caller_credential)
-        .map_err(|_error| BrokerHostError::AssociationSetup)?;
-
-    let result = serve_request_loop(core, channel, &association);
-    core.close_association(association);
-    result
-}
-
-fn serve_request_loop<T>(
-    core: &mut BrokerCore,
-    channel: &mut T,
-    association: &BrokerAssociation,
-) -> HostResult<ConnectionTermination, T::Error>
-where
-    T: HostControlChannel,
-{
-    let mut state = ConnectionState::AwaitingNegotiation;
-    loop {
-        let Some(received) = channel.recv_request().map_err(BrokerHostError::Channel)? else {
-            break;
-        };
-
-        let dispatch = handle_received_request(core, association, &mut state, received);
-        channel
-            .send_response(&dispatch.response)
-            .map_err(BrokerHostError::Channel)?;
-        if let DispatchOutcome::Close(reason) = dispatch.outcome {
-            return Ok(ConnectionTermination::BrokerClosed(reason));
-        }
-    }
-
-    Ok(ConnectionTermination::PeerClosed)
-}
-
-fn caller_credential_from_peer(
-    peer_credential: PeerCredential,
-) -> core::result::Result<CallerCredential, ()> {
-    if peer_credential == PeerCredential::Unauthenticated {
-        Ok(CallerCredential::Unauthenticated)
-    } else {
-        Err(())
-    }
-}
-
-fn handle_received_request(
-    core: &mut BrokerCore,
-    association: &BrokerAssociation,
-    state: &mut ConnectionState,
-    received: ReceivedBrokerRequest,
-) -> BrokerDispatch {
-    match received {
-        ReceivedBrokerRequest::Request(request) => {
-            handle_request(core, association, state, request)
-        }
-        _ => handle_unknown_request(*state),
-    }
-}
-
-fn handle_request(
-    core: &mut BrokerCore,
-    association: &BrokerAssociation,
-    state: &mut ConnectionState,
-    request: BrokerRequest,
-) -> BrokerDispatch {
-    match *state {
-        ConnectionState::AwaitingNegotiation => match request {
-            BrokerRequest::Negotiate { protocol_version } => {
-                negotiate_version(state, protocol_version)
-            }
-            _ => BrokerDispatch::close_after(
-                BrokerResponse::Error(ErrorCode::ProtocolState),
-                CloseReason::ProtocolViolation,
-            ),
-        },
-        ConnectionState::Active {
-            negotiated_protocol_version,
-        } => handle_active_request(core, association, negotiated_protocol_version, request),
-    }
-}
-
-fn handle_active_request(
-    core: &mut BrokerCore,
-    association: &BrokerAssociation,
-    _negotiated_protocol_version: ProtocolVersion,
-    request: BrokerRequest,
-) -> BrokerDispatch {
-    match request {
-        BrokerRequest::Negotiate { .. } => BrokerDispatch::close_after(
-            BrokerResponse::Error(ErrorCode::ProtocolState),
-            CloseReason::ProtocolViolation,
-        ),
-        BrokerRequest::Core(request) => {
-            BrokerDispatch::continue_after(handle_core_request(core, association, request))
-        }
-        _ => BrokerDispatch::continue_after(BrokerResponse::Error(ErrorCode::UnsupportedOperation)),
-    }
-}
-
-fn handle_core_request(
-    core: &mut BrokerCore,
-    association: &BrokerAssociation,
-    request: CoreRequest,
-) -> BrokerResponse {
-    match request {
-        CoreRequest::Event(request) => handle_event_request(core, association, request),
-        _ => BrokerResponse::Error(ErrorCode::UnsupportedOperation),
-    }
-}
-
-fn handle_event_request(
-    core: &mut BrokerCore,
-    association: &BrokerAssociation,
-    request: EventRequest,
-) -> BrokerResponse {
-    match request {
-        EventRequest::Create(request) => handle_core_result(
-            core.create_event_with_count(association, request.initial_count),
-            |handle| event_response(EventResponse::Create(CreateEventResponse::new(handle))),
-        ),
-        EventRequest::Wait(request) => {
-            handle_core_result(core.wait_event(association, request.handle), |outcome| {
-                event_response(EventResponse::Wait(WaitEventResponse::new(outcome)))
-            })
-        }
-        EventRequest::Add(request) => handle_core_result(
-            core.add_event(association, request.handle, request.value),
-            |readiness| event_response(EventResponse::Add(AddEventResponse::new(readiness))),
-        ),
-        EventRequest::Consume(request) => handle_core_result(
-            core.consume_event(association, request.handle, request.mode),
-            |consumption| event_response(EventResponse::Consume(consumption)),
-        ),
-        _ => BrokerResponse::Error(ErrorCode::UnsupportedOperation),
-    }
-}
-
-fn handle_unknown_request(state: ConnectionState) -> BrokerDispatch {
-    if state == ConnectionState::AwaitingNegotiation {
-        BrokerDispatch::close_after(
-            BrokerResponse::Error(ErrorCode::ProtocolState),
-            CloseReason::ProtocolViolation,
-        )
-    } else {
-        BrokerDispatch::continue_after(BrokerResponse::Error(ErrorCode::UnsupportedOperation))
-    }
-}
-
-fn negotiate_version(
-    state: &mut ConnectionState,
-    protocol_version: ProtocolVersion,
-) -> BrokerDispatch {
-    if protocol_version.is_supported_by(HOST_PROTOCOL_VERSION) {
-        *state = ConnectionState::Active {
-            negotiated_protocol_version: protocol_version,
-        };
-        BrokerDispatch::continue_after(BrokerResponse::Negotiated {
-            broker_protocol_version: HOST_PROTOCOL_VERSION,
-        })
-    } else {
-        BrokerDispatch::continue_after(BrokerResponse::VersionMismatch {
-            broker_protocol_version: HOST_PROTOCOL_VERSION,
-        })
-    }
-}
-
-fn handle_core_result<T>(
-    result: litebox_broker_core::Result<T>,
-    into_response: impl FnOnce(T) -> BrokerResponse,
-) -> BrokerResponse {
-    match result {
-        Ok(value) => into_response(value),
-        Err(error) => BrokerResponse::Error(to_protocol_error(error)),
-    }
-}
-
-const fn event_response(response: EventResponse) -> BrokerResponse {
-    BrokerResponse::Core(CoreResponse::Event(response))
-}
-
-fn to_protocol_error(error: BrokerError) -> ErrorCode {
-    match error {
-        BrokerError::PolicyDenied => ErrorCode::PolicyDenied,
-        BrokerError::UnknownObject => ErrorCode::UnknownObject,
-        BrokerError::StaleHandle => ErrorCode::StaleHandle,
-        BrokerError::WrongObjectType => ErrorCode::WrongObjectType,
-        BrokerError::InvalidRights => ErrorCode::InvalidRights,
-        BrokerError::ResourceExhausted => ErrorCode::ResourceExhausted,
-        BrokerError::WouldBlock => ErrorCode::WouldBlock,
-        BrokerError::UnsupportedOperation => ErrorCode::UnsupportedOperation,
-        _ => ErrorCode::Internal,
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum ConnectionState {
-    AwaitingNegotiation,
-    Active {
-        negotiated_protocol_version: ProtocolVersion,
-    },
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct BrokerDispatch {
-    response: BrokerResponse,
-    outcome: DispatchOutcome,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum DispatchOutcome {
-    Continue,
-    Close(CloseReason),
-}
-
-impl BrokerDispatch {
-    const fn continue_after(response: BrokerResponse) -> Self {
-        Self {
-            response,
-            outcome: DispatchOutcome::Continue,
-        }
-    }
-
-    const fn close_after(response: BrokerResponse, reason: CloseReason) -> Self {
-        Self {
-            response,
-            outcome: DispatchOutcome::Close(reason),
-        }
-    }
-}
-
-/// Reason the broker host closed the connection after sending a response.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum CloseReason {
-    /// The peer violated the request sequencing state machine.
-    ProtocolViolation,
-}
-
-impl fmt::Display for CloseReason {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::ProtocolViolation => f.write_str("protocol violation"),
-        }
-    }
-}
-
-/// Terminal outcome for a successfully served broker connection.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum ConnectionTermination {
-    /// The peer cleanly closed the channel.
-    PeerClosed,
-    /// The host sent a terminal protocol response and closed the connection.
-    BrokerClosed(CloseReason),
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use litebox_broker_core::PolicyEngine;
-    use litebox_broker_protocol::CreateEventRequest;
-
-    #[test]
-    fn host_request_handling_uses_one_broker_core() {
-        let mut core = BrokerCore::new(PolicyEngine::event_only()).unwrap();
-
-        dispatch_enforces_negotiation_state(&mut core);
-        dispatch_rejects_unsupported_protocol_version_without_activation(&mut core);
-        dispatch_handles_unknown_wire_requests_by_state(&mut core);
-        serve_connection_negotiates_routes_one_request_and_returns_peer_closed(&mut core);
-        serve_connection_closes_after_protocol_violation(&mut core);
-        serve_connection_returns_channel_error_when_response_send_fails(&mut core);
-    }
-
-    fn dispatch_enforces_negotiation_state(core: &mut BrokerCore) {
-        {
-            let (association, mut state) = new_association(core);
-
-            let dispatch = handle_request(core, &association, &mut state, event_create_request(0));
-            assert_protocol_violation(dispatch);
-            assert_eq!(state, ConnectionState::AwaitingNegotiation);
-            core.close_association(association);
-        }
-
-        {
-            let (association, mut state) = new_association(core);
-            negotiate(core, &association, &mut state);
-
-            let dispatch = handle_request(
-                core,
-                &association,
-                &mut state,
-                BrokerRequest::Negotiate {
-                    protocol_version: HOST_PROTOCOL_VERSION,
-                },
-            );
-            assert_protocol_violation(dispatch);
-            core.close_association(association);
-        }
-    }
-
-    fn dispatch_rejects_unsupported_protocol_version_without_activation(core: &mut BrokerCore) {
-        let (association, mut state) = new_association(core);
-
-        let dispatch = handle_request(
-            core,
-            &association,
-            &mut state,
-            BrokerRequest::Negotiate {
-                protocol_version: ProtocolVersion::new(HOST_PROTOCOL_VERSION.major + 1, 0),
-            },
-        );
-
-        assert_eq!(
-            dispatch.response,
-            BrokerResponse::VersionMismatch {
-                broker_protocol_version: HOST_PROTOCOL_VERSION
-            }
-        );
-        assert_eq!(dispatch.outcome, DispatchOutcome::Continue);
-        assert_eq!(state, ConnectionState::AwaitingNegotiation);
-        core.close_association(association);
-    }
-
-    fn dispatch_handles_unknown_wire_requests_by_state(core: &mut BrokerCore) {
-        {
-            let (association, mut state) = new_association(core);
-
-            let dispatch = handle_received_request(
-                core,
-                &association,
-                &mut state,
-                ReceivedBrokerRequest::Unknown,
-            );
-            assert_protocol_violation(dispatch);
-            core.close_association(association);
-        }
-
-        {
-            let (association, mut state) = new_association(core);
-            negotiate(core, &association, &mut state);
-
-            let dispatch = handle_received_request(
-                core,
-                &association,
-                &mut state,
-                ReceivedBrokerRequest::Unknown,
-            );
-
-            assert_eq!(
-                dispatch.response,
-                BrokerResponse::Error(ErrorCode::UnsupportedOperation)
-            );
-            assert_eq!(dispatch.outcome, DispatchOutcome::Continue);
-            core.close_association(association);
-        }
-    }
-
-    fn serve_connection_negotiates_routes_one_request_and_returns_peer_closed(
-        core: &mut BrokerCore,
-    ) {
-        let mut channel = FakeHostControlChannel::new(std::vec::Vec::from([
-            Ok(Some(ReceivedBrokerRequest::Request(
-                BrokerRequest::Negotiate {
-                    protocol_version: HOST_PROTOCOL_VERSION,
-                },
-            ))),
-            Ok(Some(ReceivedBrokerRequest::Request(event_create_request(
-                0,
-            )))),
-            Ok(None),
-        ]));
-
-        assert_eq!(
-            serve_connection(core, &mut channel).unwrap(),
-            ConnectionTermination::PeerClosed
-        );
-        assert_eq!(
-            channel.responses[0],
-            BrokerResponse::Negotiated {
-                broker_protocol_version: HOST_PROTOCOL_VERSION
-            }
-        );
-        let handle = match &channel.responses[1] {
-            BrokerResponse::Core(CoreResponse::Event(EventResponse::Create(response))) => {
-                response.handle
-            }
-            response => panic!("unexpected response: {response:?}"),
-        };
-        assert_ne!(handle.reference_id.get(), 0);
-    }
-
-    fn serve_connection_closes_after_protocol_violation(core: &mut BrokerCore) {
-        let mut channel = FakeHostControlChannel::new(std::vec::Vec::from([
-            Ok(Some(ReceivedBrokerRequest::Request(event_create_request(
-                0,
-            )))),
-            Ok(Some(ReceivedBrokerRequest::Request(
-                BrokerRequest::Negotiate {
-                    protocol_version: HOST_PROTOCOL_VERSION,
-                },
-            ))),
-        ]));
-
-        assert_eq!(
-            serve_connection(core, &mut channel).unwrap(),
-            ConnectionTermination::BrokerClosed(CloseReason::ProtocolViolation)
-        );
-        assert_eq!(
-            channel.responses,
-            [BrokerResponse::Error(ErrorCode::ProtocolState)]
-        );
-        assert_eq!(channel.requests.len(), 1);
-    }
-
-    fn serve_connection_returns_channel_error_when_response_send_fails(core: &mut BrokerCore) {
-        let mut channel = FakeHostControlChannel::new(std::vec::Vec::from([Ok(Some(
-            ReceivedBrokerRequest::Request(BrokerRequest::Negotiate {
-                protocol_version: HOST_PROTOCOL_VERSION,
-            }),
-        ))]));
-        channel.send_error = Some(FakeChannelError::Send);
-
-        match serve_connection(core, &mut channel) {
-            Err(BrokerHostError::Channel(FakeChannelError::Send)) => {}
-            result => panic!("unexpected serve result: {result:?}"),
-        }
-        assert!(channel.responses.is_empty());
-    }
-
-    fn assert_protocol_violation(dispatch: BrokerDispatch) {
-        assert_eq!(
-            dispatch.response,
-            BrokerResponse::Error(ErrorCode::ProtocolState)
-        );
-        assert_eq!(
-            dispatch.outcome,
-            DispatchOutcome::Close(CloseReason::ProtocolViolation)
-        );
-    }
-
-    fn new_association(core: &mut BrokerCore) -> (BrokerAssociation, ConnectionState) {
-        let association = core
-            .create_association(CallerCredential::Unauthenticated)
-            .unwrap();
-        (association, ConnectionState::AwaitingNegotiation)
-    }
-
-    fn negotiate(
-        core: &mut BrokerCore,
-        association: &BrokerAssociation,
-        state: &mut ConnectionState,
-    ) {
-        let dispatch = handle_request(
-            core,
-            association,
-            state,
-            BrokerRequest::Negotiate {
-                protocol_version: HOST_PROTOCOL_VERSION,
-            },
-        );
-        assert_eq!(
-            dispatch.response,
-            BrokerResponse::Negotiated {
-                broker_protocol_version: HOST_PROTOCOL_VERSION
-            }
-        );
-        assert_eq!(
-            *state,
-            ConnectionState::Active {
-                negotiated_protocol_version: HOST_PROTOCOL_VERSION
-            }
-        );
-    }
-
-    const fn event_request(request: EventRequest) -> BrokerRequest {
-        BrokerRequest::Core(CoreRequest::Event(request))
-    }
-
-    const fn event_create_request(initial_count: u64) -> BrokerRequest {
-        event_request(EventRequest::Create(CreateEventRequest::new(initial_count)))
-    }
-
-    #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-    enum FakeChannelError {
-        Send,
-    }
-
-    impl fmt::Display for FakeChannelError {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            match self {
-                Self::Send => f.write_str("fake send error"),
-            }
-        }
-    }
-
-    impl core::error::Error for FakeChannelError {}
-
-    struct FakeHostControlChannel {
-        requests:
-            std::vec::Vec<core::result::Result<Option<ReceivedBrokerRequest>, FakeChannelError>>,
-        responses: std::vec::Vec<BrokerResponse>,
-        send_error: Option<FakeChannelError>,
-    }
-
-    impl FakeHostControlChannel {
-        fn new(
-            requests: std::vec::Vec<
-                core::result::Result<Option<ReceivedBrokerRequest>, FakeChannelError>,
-            >,
-        ) -> Self {
-            Self {
-                requests,
-                responses: std::vec::Vec::new(),
-                send_error: None,
-            }
-        }
-    }
-
-    impl HostControlChannel for FakeHostControlChannel {
-        type Error = FakeChannelError;
-
-        fn peer_credential(&self) -> core::result::Result<PeerCredential, Self::Error> {
-            Ok(PeerCredential::Unauthenticated)
-        }
-
-        fn recv_request(
-            &mut self,
-        ) -> core::result::Result<Option<ReceivedBrokerRequest>, Self::Error> {
-            if self.requests.is_empty() {
-                Ok(None)
-            } else {
-                self.requests.remove(0)
-            }
-        }
-
-        fn send_response(
-            &mut self,
-            response: &BrokerResponse,
-        ) -> core::result::Result<(), Self::Error> {
-            if let Some(error) = self.send_error {
-                return Err(error);
-            }
-            self.responses.push(response.clone());
-            Ok(())
-        }
-    }
-}
