+++ litebox_broker_local/src/lib.rs
@@ -14,7 +14,6 @@ extern crate std;
 
 mod error;
 mod event;
-mod negotiate;
 
 use litebox_broker_protocol::{
     BrokerRequest, BrokerResponse, LocalControlChannel, ProtocolVersion, ReceivedBrokerResponse,
@@ -55,6 +54,58 @@ impl<T> BrokerLocal<T> {
 }
 
 impl<T: LocalControlChannel> BrokerLocal<T> {
+    /// Negotiates the default broker-local protocol version.
+    ///
+    /// Returns the effective protocol version this connection will speak.
+    pub fn negotiate(&mut self) -> Result<ProtocolVersion, T::Error> {
+        self.negotiate_version(LOCAL_PROTOCOL_VERSION)
+    }
+
+    /// Negotiates a caller-selected protocol version.
+    ///
+    /// Returns the effective protocol version this connection will speak. Feature
+    /// gating must use this effective version, not the broker's max-supported
+    /// version returned by the wire negotiation response.
+    pub fn negotiate_version(
+        &mut self,
+        protocol_version: ProtocolVersion,
+    ) -> Result<ProtocolVersion, T::Error> {
+        if self.state != ConnectionState::AwaitingNegotiation {
+            return Err(BrokerLocalError::AlreadyNegotiated);
+        }
+        if !protocol_version.is_supported_by(LOCAL_PROTOCOL_VERSION) {
+            return Err(BrokerLocalError::UnsupportedLocalVersion {
+                requested: protocol_version,
+                local_protocol_version: LOCAL_PROTOCOL_VERSION,
+            });
+        }
+
+        let response = self.request(BrokerRequest::Negotiate { protocol_version })?;
+        match response {
+            BrokerResponse::Negotiated {
+                broker_protocol_version,
+            } => {
+                if !protocol_version.is_supported_by(broker_protocol_version) {
+                    return Err(BrokerLocalError::IncompatibleNegotiation {
+                        requested: protocol_version,
+                        broker_protocol_version,
+                    });
+                }
+                self.state = ConnectionState::Active {
+                    negotiated_protocol_version: protocol_version,
+                };
+                Ok(protocol_version)
+            }
+            BrokerResponse::VersionMismatch {
+                broker_protocol_version,
+            } => Err(BrokerLocalError::UnsupportedVersion {
+                requested: protocol_version,
+                broker_protocol_version,
+            }),
+            response => Err(BrokerLocalError::UnexpectedResponse(response)),
+        }
+    }
+
     /// Returns the effective protocol version this connection negotiated.
     ///
     /// Feature gating must use this effective version because the broker may
